small changes to album and rating

+++ Backend/backend/springboot_example/src/main/java/onetomany/Album/AlbumRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbumRepository extends JpaRepository<Album, Long> {
+public interface AlbumRepository extends JpaRepository<Album, Integer> { // Changed from String to Integer
     // Additional custom query methods (if needed) can be added here
 }
